Creating a default gson instance
